requirements.txt
numpy
opencv-python
tensorboard
tqdm
pycocotools
matplotlib
PyYAML
ultralytics